requirements.txt
PyNaCl
six
